fix: separate get-dump and make-dump files
